+++ CapaDatos/Models/ViewModel/VM_PDV_VENTA.cs
@@ -15,6 +15,6 @@ namespace CapaDatos.Models.ViewModel
         public string Sucursal { get; set; }
         public string Vendedor { get; set; }
         public string ListaPrecios { get; set; }
-        public int VtaAccion { get; set; }
+        public Int16 VtaAccion { get; set; }
     }
 }
